// src/shared/types/cases/children.ts
export interface IChild {
  childId: string;
  fullName: string;
  firstName: string;
  middleName: string;
  lastName: string;
  birthDate?: any;
  age?: number;
  sex: string;
  eyeColor: string;
  hairColor: string;
  height: string;
  heightUnit: string;
  weight: string;
  weightUnit: string;
  images: string[];
}